Update elwiz.js
A huge update to accommodate for Home Assistant auto discovery. Some breaking changes are introduced, but it will hopefully work as before for old friends. Note that the "power.yaml" file is now DEPRECATED. Instead the new "savepower.json" file holds some received meter data for recovery after short stops. A new requirement is the "simple-json-db" for this purpose.